chore: split player functions into new file
